feat: implement api to trade based on the latest best aggregated price

+++ src/main/java/com/example/crypto_trading/repository/AggregatedPriceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AggregatedPriceRepository extends JpaRepository<AggregatedPrice, Integer> {
+public interface AggregatedPriceRepository extends JpaRepository<AggregatedPrice, Long> {
     Optional<AggregatedPrice> findByPair(String pair);
 }
